pagging and sorting

+++ springboot-project/src/main/java/com/example/demo/service/ServiceImpl.java
@@ -1,12 +1,15 @@
 package com.example.demo.service;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.PathVariable;
 
+import com.example.demo.exception.IdNotFoundException;
+import com.example.demo.exception.NameNotFoundException;
 import com.example.demo.model.Company;
 import com.example.demo.model.Employee;
 import com.example.demo.repository.CompanyRepository;
@@ -27,8 +30,16 @@ public class ServiceImpl implements CompanyService {
 	}
 
 	@Override
-	public Company findById(int empid) {
-		return repository.findById(empid).get();
+	public Company findById(int empid) throws IdNotFoundException  {
+		
+		if(employeeRepository.existsById(empid))
+		{
+			return repository.findById(empid).get();
+
+		}
+		else {
+			throw new IdNotFoundException();
+		}
 	}
 
 	@Override
@@ -41,24 +52,33 @@ public class ServiceImpl implements CompanyService {
 	}
 
 	@Override
-	public String deleteById(int empid) {
-
-		  employeeRepository.deleteById(empid);
-		  return"deleted";
-	}
-
-	
-
-	
-
-	@Override
-	public String findByName(String empname) {
+	public String deleteById(int empid) throws IdNotFoundException {
 		
-	employeeRepository.findByempname(empname);
-		return "get";
+		if(employeeRepository.existsById(empid))
+		{
 
-
-					
+		   employeeRepository.deleteById(empid);
+		}
+		else
+		{
+			throw new IdNotFoundException();
+		}
+		
+		return "";
+		  
+	}
+	
+	@Override
+	public String findByName(String empname) throws NameNotFoundException {
+		if(employeeRepository.equals(empname))
+		{
+			return employeeRepository.findByempname(empname);
+		}
+		else
+		{
+			throw new NameNotFoundException();
+		}
+							
 	}
 
 	@Override
@@ -67,6 +87,27 @@ public class ServiceImpl implements CompanyService {
 		return employeeRepository.findAll();
 	}
 
+	@Override
+	public List<Employee> findEmployeeWithSorting(String field) {
+		
+		return employeeRepository.findAll(Sort.by(Sort.Direction.ASC,field));
+	}
+
+	@Override
+	public Page<Employee> findEmployeeWithPagination(int offset, int pagesize) {
+		 Page<Employee> findAll = employeeRepository.findAll(PageRequest.of(offset, pagesize));
+		return findAll;
+		
+	}
+
+	@Override
+	public Page<Employee> findEmployeeWithPaginationAndSort(int offset, int pagesize, String field) {
+		
+		Page<Employee> findAll1=employeeRepository.findAll(PageRequest.of(offset, pagesize).withSort(Sort.by(field)));
+		return findAll1;
+	}
+
+
